+++ client/src/main.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App.tsx";
+import App from "./temp/App.tsx";
 import "./index.css";
 
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
